Skills/Stock/HongKong_Stock/strategy-optimizer/requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
scikit-learn>=1.2.0
